ajout pawn captures masks

+++ lib/bitboard.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <stdio.h>
 
-uint64_t create_1bit_board(uint64_t exponent);
+uint64_t create_1bit_board(int exponent);
 void print_bitboard(uint64_t bitboard);
 
 #endif
